+++ examples/expo/src/App.tsx
@@ -7,6 +7,7 @@ import { CardExample } from "@/src/screens/Card";
 import { ThreeDSecureExample } from "@/src/screens/ThreeDSecure";
 import { Image, ImageSourcePropType } from "react-native";
 import { EncryptExample } from "@/src/screens/Encrypt";
+import { PayExample } from "./screens/Pay";
 
 function tabBarIcon(source: ImageSourcePropType) {
   return function TabBarIcon({ color, size }: { color: string; size: number }) {
@@ -27,6 +28,17 @@ const Tabs = createBottomTabNavigator({
       options: {
         headerShown: false,
         title: "Card",
+        tabBarIcon: tabBarIcon(
+          require("@/assets/images/icons/credit-card.png")
+        ),
+      },
+    },
+
+    Pay: {
+      screen: PayExample,
+      options: {
+        headerShown: false,
+        title: "Pay",
         tabBarIcon: tabBarIcon(
           require("@/assets/images/icons/credit-card.png")
         ),
