+++ packages/react-native-v2/src/ApplePayButton/index.tsx
@@ -1,0 +1,4 @@
+import { EvervaultPaymentViewProps } from './types';
+
+export * from './types';
+export const ApplePayButton: React.FC<EvervaultPaymentViewProps> = () => null;
