+++ packages/react-native-v2/src/example-ios-usage.tsx
@@ -1,128 +1,128 @@
-import React, { useState } from 'react';
-import { View, Text, StyleSheet, Alert } from 'react-native';
-import { EvervaultPaymentView } from './EvervaultPaymentView.android';
-import { Config, Transaction } from './EvervaultPaymentViewNativeComponent.android';
+// import React, { useState } from 'react';
+// import { View, Text, StyleSheet, Alert } from 'react-native';
+// import { EvervaultPaymentView } from './EvervaultPaymentView.android';
+// import { Config, Transaction } from './EvervaultPaymentViewNativeComponent.android';
 
-const ExampleIOSUsage: React.FC = () => {
-  const [isProcessing, setIsProcessing] = useState(false);
+// const ExampleIOSUsage: React.FC = () => {
+//   const [isProcessing, setIsProcessing] = useState(false);
 
-  // Configuration for Evervault
-  const config: Config = {
-    appId: 'your-evervault-app-id',
-    merchantId: 'your-apple-merchant-id'
-  };
+//   // Configuration for Evervault
+//   const config: Config = {
+//     appId: 'your-evervault-app-id',
+//     merchantId: 'your-apple-merchant-id'
+//   };
 
-  // Transaction details
-  const transaction: Transaction = {
-    total: '29.99',
-    currency: 'USD',
-    country: 'US',
-    lineItems: [
-      {
-        label: 'Premium Subscription',
-        amount: '29.99'
-      }
-    ]
-  };
+//   // Transaction details
+//   const transaction: Transaction = {
+//     total: '29.99',
+//     currency: 'USD',
+//     country: 'US',
+//     lineItems: [
+//       {
+//         label: 'Premium Subscription',
+//         amount: '29.99'
+//       }
+//     ]
+//   };
 
-  const handleDidAuthorizePayment = (data: any) => {
-    console.log('Payment authorized:', data);
-    setIsProcessing(false);
+//   const handleDidAuthorizePayment = (data: any) => {
+//     console.log('Payment authorized:', data);
+//     setIsProcessing(false);
     
-    // Handle the payment response
-    // The data contains the encrypted payment token that you can send to your backend
-    Alert.alert(
-      'Payment Successful',
-      'Your payment has been processed successfully!'
-    );
-  };
+//     // Handle the payment response
+//     // The data contains the encrypted payment token that you can send to your backend
+//     Alert.alert(
+//       'Payment Successful',
+//       'Your payment has been processed successfully!'
+//     );
+//   };
 
-  const handleDidFinishWithResult = (data: { success: boolean; error?: string }) => {
-    console.log('Payment finished:', data);
-    setIsProcessing(false);
+//   const handleDidFinishWithResult = (data: { success: boolean; error?: string }) => {
+//     console.log('Payment finished:', data);
+//     setIsProcessing(false);
     
-    if (!data.success) {
-      Alert.alert(
-        'Payment Failed',
-        data.error || 'An error occurred during payment processing.'
-      );
-    }
-  };
+//     if (!data.success) {
+//       Alert.alert(
+//         'Payment Failed',
+//         data.error || 'An error occurred during payment processing.'
+//       );
+//     }
+//   };
 
-  const handlePrepareTransaction = () => {
-    console.log('Preparing transaction...');
-    // You can modify the transaction here if needed
-  };
+//   const handlePrepareTransaction = () => {
+//     console.log('Preparing transaction...');
+//     // You can modify the transaction here if needed
+//   };
 
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Evervault Apple Pay Example</Text>
-      <Text style={styles.subtitle}>iOS Integration</Text>
+//   return (
+//     <View style={styles.container}>
+//       <Text style={styles.title}>Evervault Apple Pay Example</Text>
+//       <Text style={styles.subtitle}>iOS Integration</Text>
       
-      <View style={styles.paymentContainer}>
-        <EvervaultPaymentView
-          config={config}
-          transaction={transaction}
-          buttonType="buy"
-          buttonTheme="automatic"
-          borderRadius={8}
-          allowedCardNetworks={['VISA', 'MASTERCARD', 'AMEX']}
-          onDidAuthorizePayment={handleDidAuthorizePayment}
-          onDidFinishWithResult={handleDidFinishWithResult}
-          onPrepareTransaction={handlePrepareTransaction}
-          style={styles.paymentButton}
-        />
-      </View>
+//       <View style={styles.paymentContainer}>
+//         <EvervaultPaymentView
+//           config={config}
+//           transaction={transaction}
+//           buttonType="buy"
+//           buttonTheme="automatic"
+//           borderRadius={8}
+//           allowedCardNetworks={['VISA', 'MASTERCARD', 'AMEX']}
+//           onDidAuthorizePayment={handleDidAuthorizePayment}
+//           onDidFinishWithResult={handleDidFinishWithResult}
+//           onPrepareTransaction={handlePrepareTransaction}
+//           style={styles.paymentButton}
+//         />
+//       </View>
       
-      {isProcessing && (
-        <Text style={styles.processingText}>Processing payment...</Text>
-      )}
+//       {isProcessing && (
+//         <Text style={styles.processingText}>Processing payment...</Text>
+//       )}
       
-      <Text style={styles.note}>
-        Note: This example requires a valid Apple Merchant ID and Evervault App ID.
-        Make sure Apple Pay is enabled on your device.
-      </Text>
-    </View>
-  );
-};
+//       <Text style={styles.note}>
+//         Note: This example requires a valid Apple Merchant ID and Evervault App ID.
+//         Make sure Apple Pay is enabled on your device.
+//       </Text>
+//     </View>
+//   );
+// };
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 20,
-    backgroundColor: '#f5f5f5',
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    textAlign: 'center',
-    marginBottom: 8,
-  },
-  subtitle: {
-    fontSize: 16,
-    textAlign: 'center',
-    color: '#666',
-    marginBottom: 30,
-  },
-  paymentContainer: {
-    alignItems: 'center',
-    marginBottom: 20,
-  },
-  paymentButton: {
-    width: '100%',
-    height: 50,
-  },
-  processingText: {
-    textAlign: 'center',
-    color: '#007AFF',
-    marginBottom: 20,
-  },
-  note: {
-    fontSize: 12,
-    color: '#999',
-    textAlign: 'center',
-    fontStyle: 'italic',
-  },
-});
+// const styles = StyleSheet.create({
+//   container: {
+//     flex: 1,
+//     padding: 20,
+//     backgroundColor: '#f5f5f5',
+//   },
+//   title: {
+//     fontSize: 24,
+//     fontWeight: 'bold',
+//     textAlign: 'center',
+//     marginBottom: 8,
+//   },
+//   subtitle: {
+//     fontSize: 16,
+//     textAlign: 'center',
+//     color: '#666',
+//     marginBottom: 30,
+//   },
+//   paymentContainer: {
+//     alignItems: 'center',
+//     marginBottom: 20,
+//   },
+//   paymentButton: {
+//     width: '100%',
+//     height: 50,
+//   },
+//   processingText: {
+//     textAlign: 'center',
+//     color: '#007AFF',
+//     marginBottom: 20,
+//   },
+//   note: {
+//     fontSize: 12,
+//     color: '#999',
+//     textAlign: 'center',
+//     fontStyle: 'italic',
+//   },
+// });
 
-export default ExampleIOSUsage; 
+// export default ExampleIOSUsage;
